Refactor: DB 구조 변경에 따른 수강 이력 로직 변경
