--- src/main/kotlin/com/hhplus/enrollment/lecture/application/info/LectureHistoryCommandInfo.kt
@@ -1,9 +1,0 @@
-package com.hhplus.enrollment.lecture.application.info
-
-import com.hhplus.enrollment.lecture.domain.data.LectureHistoryQueryData
-
-data class LectureHistoryCommandInfo(
-    var userId: Long,
-) {
-    fun toData() = LectureHistoryQueryData(userId = userId)
-}
