+++ src/main/kotlin/com/hhplus/enrollment/lecture/domain/data/LectureHistoryQueryData.kt
@@ -3,11 +3,7 @@ package com.hhplus.enrollment.lecture.domain.data
 import com.hhplus.enrollment.lecture.infrastructure.dto.LectureHistoryQueryDto
 
 data class LectureHistoryQueryData(
-    var userId: Long
+    var traineeId: Long
 ) {
-    init {
-        require(userId >= 0) { "User ID는 0 이상이어야 합니다."}
-    }
-
-    fun toDto(): LectureHistoryQueryDto = LectureHistoryQueryDto(userId)
+    fun toDto(): LectureHistoryQueryDto = LectureHistoryQueryDto(traineeId)
 }
